# openapi/generated_server/go/model_product.go
/*
 * Fridge API
 *
 * OpenAPI 3.0.0 study - a fridge API
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

// Product - Single product
type Product struct {

	Name string `json:"name,omitempty"`

	Quantity float32 `json:"quantity,omitempty"`
}